Updated CLI Commands to api-mesh:*
