+++ src/commands/api-mesh/__tests__/delete.test.js
@@ -10,9 +10,11 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-const mockConsoleCLIInstance = {};
 jest.mock('@adobe/aio-lib-env');
 jest.mock('@adobe/aio-cli-lib-console');
+
+const mockConsoleCLIInstance = {};
+
 const orgs = [{ id: '1234', code: 'CODE1234@AdobeOrg', name: 'ORG01', type: 'entp' }];
 const selectedOrg = { id: '1234', code: 'CODE1234@AdobeOrg', name: 'ORG01', type: 'entp' };
 
@@ -41,36 +43,40 @@ jest.mock('@adobe/aio-cli-lib-console', () => ({
 	cleanStdOut: jest.fn(),
 }));
 jest.mock('@adobe/aio-lib-ims');
-const GetCommand = require('../get');
-const { SchemaServiceClient } = require('../../../../classes/SchemaServiceClient');
-const mockGetMesh = require('../../../__fixtures__/sample_mesh.json');
 
-describe('get command tests', () => {
+const DeleteCommand = require('../delete');
+const { SchemaServiceClient } = require('../../../classes/SchemaServiceClient');
+
+describe('delete command tests', () => {
 	beforeEach(() => {
 		setDefaultMockConsoleCLI();
 	});
+
 	afterEach(() => {
 		jest.restoreAllMocks();
 	});
 
-	test('get-mesh-missing-meshId', async () => {
-		const runResult = GetCommand.run([]);
+	test('delete-mesh-missing-meshId', async () => {
+		const runResult = DeleteCommand.run([]);
 
 		return runResult.catch(err => {
 			expect(err).toHaveProperty(
 				'message',
 				expect.stringMatching(
-					/^Unable to get mesh\. Please check the details and try again\. If the error persists please contact support\. RequestId: [a-z A-Z 0-9 -_]+/,
+					/^Unable to delete mesh\. Please check the details and try again\. If the error persists please contact support\. RequestId: [a-z A-Z 0-9 -_]+/,
 				),
 			);
 		});
 	});
 
-	test('get-mesh-with-meshId', async () => {
-		jest.spyOn(SchemaServiceClient.prototype, 'getMesh').mockImplementation(meshId => mockGetMesh);
+	test('delete-mesh-with-meshId', async () => {
+		jest
+			.spyOn(SchemaServiceClient.prototype, 'deleteMesh')
+			.mockImplementation(() => Promise.resolve({}));
 		expect.assertions(1);
 		const meshId = 'sample_merchant';
-		const runResult = GetCommand.run([meshId]);
-		await expect(runResult).resolves.toEqual(mockGetMesh);
+		const runResult = DeleteCommand.run([meshId]);
+
+		await expect(runResult).resolves.toEqual({});
 	});
 });
