+++ src/helpers.js
@@ -32,7 +32,7 @@ const CONSOLE_API_KEYS = {
  * @returns {any} Returns a config object or null
  */
 async function getCommerceAdminConfig() {
-	const configFile = Config.get('mesh-api-manager.configPath');
+	const configFile = Config.get('api-mesh.configPath');
 
 	if (!configFile) {
 		return {
@@ -44,7 +44,7 @@ async function getCommerceAdminConfig() {
 		try {
 			if (!fs.existsSync(configFile)) {
 				throw new Error(
-					`Config file does not exist. Please run the command: aio config:set mesh-api-manager.configPath <path_to_json_file> with a valid file.`,
+					`Config file does not exist. Please run the command: aio config:set api-mesh.configPath <path_to_json_file> with a valid file.`,
 				);
 			}
 
@@ -67,7 +67,7 @@ async function getCommerceAdminConfig() {
 			};
 		} catch (error) {
 			logger.error(
-				'Please run the command: aio config:set mesh-api-manager.configPath <path_to_json_file> with a valid config file.',
+				'Please run the command: aio config:set api-mesh.configPath <path_to_json_file> with a valid config file.',
 			);
 
 			throw new Error(error);
